# 03_INFRASTRUCTURE/openapi.yaml
openapi: 3.0.3
info:
  title: NOVA RIFT API (Dual Mode)
  description: |
    API interna para la App móvil Nova Rift. Incluye los módulos:
    - **Live-Rift**: Análisis táctico en tiempo real durante partida.
    - **Elo Forge**: Entrenador de progreso basado en historial de métricas.
    - **Identity**: Verificación de identidad de invocador vía Riot API.
    
    ## Modelo de Datos
    Las respuestas 200 usan schemas flexibles (type: object) por compatibilidad
    con el backend dinámico. Los schemas documentales en components/schemas
    detallan los campos reales de cada respuesta.
  version: 3.0.0
servers:
  - url: http://localhost:8080
    description: Servidor de Desarrollo (Local)
paths:
...
  /api/v1/identity/summoner/{riotId}:
    get:
      operationId: getSummonerByRiotId
      summary: Obtiene datos del invocador por RiotID (path)
      description: |
        Respuesta 200: { puuid, gameName, tagLine, region, summonerLevel, profileIconId }
        Errores: { error: "mensaje descriptivo" }
      tags: [Identity]
      parameters:
        - name: riotId
          in: path
          required: true
          schema:
            type: string
          description: "Formato: NombreInvocador-TAG (guion en lugar de #)"
      responses:
        '200':
          description: OK - Datos del invocador
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Formato de RiotID inválido (falta separador -)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invocador no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Riot API rate limit excedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: API Key de Riot inválida o caducada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/identity/summoner:
    get:
      operationId: getSummonerByRiotIdQuery
      summary: Obtiene datos del invocador por RiotID (query param)
      description: Variante con query parameter, usada por el frontend para evitar problemas de encoding en el path.
      tags: [Identity]
      parameters:
        - name: riotId
          in: query
          required: true
          schema:
            type: string
          description: "Formato: NombreInvocador-TAG (guion en lugar de #)"
      responses:
        '200':
          description: OK - Datos del invocador
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Formato de RiotID inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invocador no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Mensaje de error descriptivo

    # =============================================
    # SCHEMAS DOCUMENTALES (no usados en generación de código,
    # solo como referencia de contrato para el frontend)
    # =============================================
...